[tests] Add e2e utility script
To-date, invoking the e2e test suite locally has been a tedious process, with the developer
having to keep track of and manually enter the steps necessary (Chromedriver, BINARY_PATH,
mocha, etc) to run the tests on their local machine.

This script greatly simplifies invoking the tests via the command `npm run e2e`, and also
sets the stage for running the test suite on all platforms (not just Mac) in the future.